@@ -1,79 +1,81 @@
-source 'https://rubygems.org'
+source "https://rubygems.org"
+git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-git_source(:github) do |repo_name|
-  repo_name = "#{repo_name}/#{repo_name}" unless repo_name.include?("/")
-  "https://github.com/#{repo_name}.git"
-end
+ruby "3.1.1"
 
-gem 'bundler', '~> 2.0', '>= 2.0.1'
+gem 'bundler', '~> 2.3', '>= 2.3.7'
 
-# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '~> 5.0.3'
-# Use mysql as the database for Active Record
-gem 'mysql2', '~> 0.5.2'
-# Use Puma as the app server
-gem 'puma', '~> 3.12.1'
-# Use SCSS for stylesheets
-gem 'sass-rails', '~> 5.0'
-# Use Uglifier as compressor for JavaScript assets
-gem 'uglifier', '>= 1.3.0'
-# Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 4.2'
-# See https://github.com/rails/execjs#readme for more supported runtimes
-# gem 'therubyracer', platforms: :ruby
+gem "rails", "~> 7.0.2", ">= 7.0.2.2"
 
-# Use jquery as the JavaScript library
-gem 'jquery-rails'
+# The original asset pipeline for Rails [https://github.com/rails/sprockets-rails]
+gem "sprockets-rails"
 
-# fix turbolinks jquery selector BS
-# gem 'jquery-turbolinks'
-# turbolinks was killing my sort links, I tried jquery-turbolinks with no luck... disabling
-# gem 'turbolinks', '~> 5.0.0'
+gem 'mysql2', '~> 0.5.3'
+gem "puma", "~> 5.0"
+# Use JavaScript with ESM import maps [https://github.com/rails/importmap-rails]
+gem "importmap-rails"
 
-# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.5'
+# Hotwire's SPA-like page accelerator [https://turbo.hotwired.dev]
+gem "turbo-rails"
 
-# Use ActiveModel has_secure_password
-gem 'bcrypt', '3.1.11'
+# Hotwire's modest JavaScript framework [https://stimulus.hotwired.dev]
+gem "stimulus-rails"
 
+# Use Redis adapter to run Action Cable in production
+# gem "redis", "~> 4.0"
+
+# Use Kredis to get higher-level data types in Redis [https://github.com/rails/kredis]
+# gem "kredis"
+
+# Use Active Model has_secure_password [https://guides.rubyonrails.org/active_model_basics.html#securepassword]
+# gem "bcrypt", "~> 3.1.7"
+
+# Windows does not include zoneinfo files, so bundle the tzinfo-data gem
+gem "tzinfo-data", platforms: %i[ mingw mswin x64_mingw jruby ]
+
+# Reduces boot times through caching; required in config/boot.rb
+gem "bootsnap", require: false
+
+# Use Sass to process CSS
+gem "sassc-rails"
+
+gem 'bcrypt', '~> 3.1', '>= 3.1.16'
+gem 'font-awesome-rails', '~> 4.7', '>= 4.7.0.8'
+gem 'bootstrap', '~> 5.1', '>= 5.1.3'
+gem 'coffee-script', '~> 2.4', '>= 2.4.1'
+gem 'jquery-rails', '~> 4.4'
+gem 'tinymce-rails'
+gem 'jbuilder', '~> 2.11', '>= 2.11.5'
 # Use Kaminari for pagination
 gem 'kaminari', '~> 0.16'
+gem 'octicons_helper'
+gem 'time_difference'
+gem 'okcomputer'
+gem 'lograge', '~> 0.10.0'
 
-gem 'sprockets-rails', :require => 'sprockets/railtie'
-gem 'bootstrap', '~> 4.1.1'
 
-# Use Font Awesome Rails for Font Awesome icons
-gem 'font-awesome-rails', '~> 4.7'
-
-gem 'coffee-script', '~> 2.4.1'
+# Use Active Storage variants [https://guides.rubyonrails.org/active_storage_overview.html#transforming-images]
+# gem "image_processing", "~> 1.2"
 
 group :development, :test do
-  # Call 'byebug' anywhere in the code to stop execution and get a debugger console
-  gem 'byebug', platform: :mri
+  # See https://guides.rubyonrails.org/debugging_rails_applications.html#debugging-with-the-debug-gem
+  gem "debug", platforms: %i[ mri mingw x64_mingw ]
 end
 
 group :development do
-  # Access an IRB console on exception pages or by using <%= console %> anywhere in the code.
-  gem 'web-console', '>= 3.3.0'
-  gem 'listen', '~> 3.0.5'
-  # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
-  gem 'spring'
-  gem 'spring-watcher-listen', '~> 2.0.0'
+  # Use console on exceptions pages [https://github.com/rails/web-console]
+  gem "web-console"
+
+  # Add speed badges [https://github.com/MiniProfiler/rack-mini-profiler]
+  # gem "rack-mini-profiler"
+
+  # Speed up commands on slow machines / big apps [https://github.com/rails/spring]
+  # gem "spring"
 end
 
-# Windows does not include zoneinfo files, so bundle the tzinfo-data gem
-gem 'tzinfo-data', platforms: [:mingw, :mswin, :x64_mingw, :jruby]
-
-gem 'minitest', '~> 5.1'
-
-gem 'tinymce-rails'
-gem 'octicons_helper'
-
-gem 'time_difference'
-
-# DB debugging... shows where queries are coming from
-# add this to an initializer: ActiveRecordQueryTrace.enabled = true
-# gem 'active_record_query_trace'
-
-gem 'okcomputer'
-gem 'lograge', '~> 0.10.0'
+group :test do
+  # Use system testing [https://guides.rubyonrails.org/testing.html#system-testing]
+  gem "capybara"
+  gem "selenium-webdriver"
+  gem "webdrivers"
+end
