@@ -14,7 +14,7 @@ class TiersController < ApplicationController
     @tier = Tier.find(params[:id])
     @error = @tier # tell _error_messages.html.erb to use this object for form errors
 
-    if @tier.update_attributes(tier_params)
+    if @tier.update(tier_params)
       logger.info("user #{current_user.id} updated tier #{@tier.id}")
       flash[:success] = 'tier updated'
       index
