@@ -1,4 +1,4 @@
-class AddPostSticky < ActiveRecord::Migration
+class AddPostSticky < ActiveRecord::Migration[5.0]
   def change
     add_column :posts, :sticky, :boolean, null: true, after: :body
   end
