@@ -1,4 +1,4 @@
-class SetStickyDefault < ActiveRecord::Migration
+class SetStickyDefault < ActiveRecord::Migration[5.0]
   def change
     change_column :posts, :sticky, :boolean, null: false, default: false, after: :body
   end
