@@ -1,4 +1,4 @@
-class CreateTiers < ActiveRecord::Migration
+class CreateTiers < ActiveRecord::Migration[5.0]
   def change
     create_table :tiers do |t|
       t.string :label, null: false, limit: 20
