@@ -1,4 +1,4 @@
-class CreateBuildingTypes < ActiveRecord::Migration
+class CreateBuildingTypes < ActiveRecord::Migration[5.0]
   def change
     create_table :building_types do |t|
       t.string :description, null: true, limit: 255
