@@ -1,4 +1,4 @@
-class CreateBuildings < ActiveRecord::Migration
+class CreateBuildings < ActiveRecord::Migration[5.0]
   def change
     create_table :buildings do |t|
       t.references :building_type, null: false
