@@ -1,4 +1,4 @@
-class AddConfirmationTimeToReservations < ActiveRecord::Migration
+class AddConfirmationTimeToReservations < ActiveRecord::Migration[5.0]
   def change
     add_column :reservations, :confirmed_time, :datetime
   end
