@@ -2,7 +2,7 @@
 // You can generate new channels where WebSocket features live using the rails
 // generate channel command.
 //
-//= require action_cable
+//= require actioncable
 //= require_self
 //= require_tree ./channels
 
