@@ -1,4 +1,4 @@
-class ReservationAddPaymentDetails < ActiveRecord::Migration[5.0]
+class ReservationAddPaymentDetails < ActiveRecord::Migration[5.0][5.0]
 
   def change
     create_table :payment_types do |t|
