@@ -1,4 +1,4 @@
-class CreateInstants < ActiveRecord::Migration[5.0]
+class CreateInstants < ActiveRecord::Migration[5.0][5.0]
   def change
     create_table :instants do |t|
       t.string :name
