@@ -1,4 +1,4 @@
-class AddDisplayOrderToPosts < ActiveRecord::Migration[5.0]
+class AddDisplayOrderToPosts < ActiveRecord::Migration[5.0][5.0]
   def change
     add_column :posts, :display_order, :integer, default: 50
   end
