@@ -1,4 +1,4 @@
-class RemoveStickyFromPosts < ActiveRecord::Migration[5.0]
+class RemoveStickyFromPosts < ActiveRecord::Migration[5.0][5.0]
   def change
       remove_column :posts, :sticky, :boolean
   end
