@@ -1,4 +1,4 @@
-class ExtendEmailBodyFields < ActiveRecord::Migration[5.0]
+class ExtendEmailBodyFields < ActiveRecord::Migration[5.0][5.0]
   def change
     change_column :tiers, :email_text, :text
     change_column :tiers, :email_html, :text
