@@ -1,4 +1,4 @@
-class DropReservationConfirmation < ActiveRecord::Migration[5.0]
+class DropReservationConfirmation < ActiveRecord::Migration[5.0][5.0]
   def change
     change_table :reservations do |t|
       t.remove :confirmed_time
